requirements/base.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
qrcode==7.4.2
tzdata==2025.2
psycopg2-binary==2.9.5
Flask-Migrate==4.1.0
requests==2.32.3
flask-restx==1.3.0
